refactor: type Container with component props
