+++ packages/epo-react-lib/src/content-blocks/SimpleTable/SimpleTable.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent, Fragment } from "react";
-import Container from "@/layout/Container";
+import Container from "@/molecules/Container";
 import * as Styled from "./styles";
 
 export interface SimpleTableRow {
